-Fix: proper usage of metaclasss between python 2 and 3.
Python3 uses a different approach of defining metaclasses, as such,
there was something going off that was causing issues.

The Singleton metaclass assumes that SoundManager is only instanced
once, and as such, does it's thread spawning and audio initialisation
inside it's __init__... now seeing python3 doesn't listen to the
__metaclasss___ property on a class, this was actually done multiple
times; causing the audio subsystem to attempt to initialize multiple
times, causing a segfault in the underlying libraries.